Trash bags are now worth 4 points each instead of 3
I talked with Alyssa yesterday, and she mentioned this. I would have
just mentioned it in the PR and had you (@stevensona) fix it, but I'm
working with Heroku support on an issue with review apps, and we wanted
to see if pushing new changes to a pull request from a non-master branch
on in a forked repo would trigger a deploy on an existing review app.