Actually it's too messy having the namespace listed as a separate class
More specifically, it's messy to have to say `:not(.admin)`. I don't
like using both dashes and hyphens but I do want to distinguish between
multiword controller names and controllers that live under a namespace.
That might not be necessary but I do think it will make it easier to
find stuff. (?)